works with 1 or 2 buffersize

+++ get_next_line.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dtome-pe <dtome-pe@student.42.fr>          +#+  +:+       +#+        */
+/*   By: theonewhoknew <theonewhoknew@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/11 11:08:18 by dtome-pe          #+#    #+#             */
-/*   Updated: 2023/05/11 13:26:31 by dtome-pe         ###   ########.fr       */
+/*   Updated: 2023/05/11 17:27:27 by theonewhokn      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strchr(const char *s, int c, int ret);
 size_t	ft_strlcpy(char *dst, const char *src,
 			size_t dstsize, unsigned int start);
 char	*ft_strdup(const char *s1, int ret);
-char	*remove_string(char *tmp, char c, int ret);
+char	*remove_string(char *tmp, char c, size_t ret);
 char	*create_string(char *tmp, char c, int ret);
 char	*create_nullstring(char *tmp, char c, int ret);
 char	*get_next_line(int fd);
